Table: moved type file
