+++ public/app/plugins/panel/table/types.ts
